main/requirements.txt
boto3==1.26.114
botocore==1.29.114
channels==3.0.4
channels-redis==3.3.1
crispy-bootstrap5==0.7
daphne==3.0.2
Django==4.0.7
django-allauth==0.54.0
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-extensions==3.2.1
django-redis==5.2.0
django-storages==1.13.2
gunicorn==20.1.0
imageio==2.28.0
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.23.5
pandas==1.5.3
playwright==1.30.0
psycopg2==2.9.3
Pillow==9.5.0
PyPDF2==3.0.1
PySocks==1.7.1
redis==4.5.4
youtube-transcript-api==0.5.0
whitenoise==6.2.0
pytube 
SpeechRecognition 
pydub
ffmpeg-python
django-mptt==0.14.0
requests==2.31.0
m3u8==3.5.0
django-login-required-middleware==0.9.0
django-ckeditor==6.6.1
django-imagekit==4.1.0


